Revert renaming to ease PR review.
